test: add inline verification helper
