+++ cmd/privdrop/main.go
@@ -33,32 +33,6 @@ func newRunner() *runner {
 	}
 }
 
-func newRunnerWithDeps(
-	ensure func(escalate.Options, *zap.Logger) (bool, error),
-	drop func(escalate.Options, *zap.Logger) error,
-	syncFunc func(*zap.Logger),
-	exitFunc func(int),
-	loggerFunc func() *zap.Logger,
-) *runner {
-	r := newRunner()
-	if ensure != nil {
-		r.ensureRootOrReexec = ensure
-	}
-	if drop != nil {
-		r.dropToInvokerIfSudo = drop
-	}
-	if syncFunc != nil {
-		r.syncLogger = syncFunc
-	}
-	if exitFunc != nil {
-		r.exit = exitFunc
-	}
-	if loggerFunc != nil {
-		r.newLogger = loggerFunc
-	}
-	return r
-}
-
 func (r *runner) run(logger *zap.Logger) int {
 	reexeced, err := r.ensureRootOrReexec(escalate.Options{}, logger)
 	if err != nil {
