+++ internal/blockio/buffer_pool.go
@@ -60,17 +60,3 @@ func putAlignedBlockBuffer(buf []byte) {
 		}
 	}
 }
-
-func purgeAlignedBlockBufferPools() {
-	bufferPools.Range(func(k, v any) bool {
-		bufferPools.Delete(k)
-		return true
-	})
-	mmappedBuffers.Range(func(k, v any) bool {
-		if b, ok := v.([]byte); ok {
-			unix.Munmap(b)
-		}
-		mmappedBuffers.Delete(k)
-		return true
-	})
-}
